requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
python-multipart==0.0.5
pillow
torch 
torchvision 
torchaudio